Update Minesweeper
Did quite a bit today, added the Get Input function to get coordinates from the user. Coordinates now also can be selected.  I made it so we have 2 field creations. The initial board which is filled with 'x'. Then each time it loops through the while(gameOver != 9) loop it will assign whatever coordinate the user put in from the reveal array to the actual board and then prints that. I don't have it fully working yet. I also started implementing a way to mark mines and a way to win, but currently you can mark an infinite amount of mines so that needs to be fixed.  I also put in system("CLR"); in certain spots just so that we dont have multiple fields in the output box. 

If you could implement the function that clears all the white space up to the next mines would be helpful and then just anything else you see that needs to be done.